api/pyproject.toml
[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "doctr-api"
version = "0.11.1a0"
description = "Backend template for your OCR API with docTR"
authors = ["Mindee <contact@mindee.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
python-doctr = {git = "https://github.com/mindee/doctr.git", extras = ['torch'], branch = "main" }
# Fastapi: minimum version required to avoid pydantic error
# cf. https://github.com/tiangolo/fastapi/issues/4168
fastapi = ">=0.73.0"
uvicorn = ">=0.11.1"
python-multipart = ">=0.0.5"
ocrmypdf = ">=16.10.1"
PyPDF2 = ">=3.0.1"

[tool.poetry.dev-dependencies]
pytest = ">=5.3.2"
pytest-asyncio = ">=0.14.0"
httpx = ">=0.23.0"
requests = ">=2.20.0"
